Rework internal I/O timeout logic
Because we only use a few methods of the AsyncReadExt/AsyncWriteExt
traits on the stream, it isn't much of a maintenance burden to just wrap
the stream in a BufStream for bidirectional buffering, and move the
timeout into this wrapper.

This also helped identify a bug where the timeout was being shared
across calls to read/write when it should have been absolute.