Reimplement Blake2b using stable SIMD intrinsics.
This implementation now only provides acceleration on x86/x86_64
architectures, but no longer relies on deprecated nightly features and
ABIs. The older implementation was actually hampering performance on
NEON enabled chips so I shall get to that along with Curve25519 when I
have the bandwidth.

With this commit, nightly is only required for avx512 optimizations in
curve25519.