Update docs to mention atomicity
Previously, the content was just an unsafe mutable static and I just
chose to spray-and-pray during races, but it turns out that parking_lot
is surprisingly more performant when there are more reads than writes,
so I decided to just use a read-write lock over the content itself.

This actual change happened *very* early on, but I only now realised
that I forgot to update the documentation. This is unfortunately going
to also look kind of ugly in blame because I did not commit the man page
that I had written at the very beginning until very recently.